+++ packages/utils/src/index.ts
@@ -1,1 +1,2 @@
+export * from "./optional-require";
 export * from "./utils";